Moves VAE to own script
